# internal/model/album.go
/*
 * Otto image service
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

import (
	"time"
)

type Album struct {

	Uuid string `json:"uuid"`

	CreatedAt time.Time `json:"created_at"`

	UpdatedAt time.Time `json:"updated_at,omitempty"`

	Link string `json:"link"`

	AlbumType AlbumType `json:"album_type"`

	Visibility Visibility `json:"visibility"`

	Name string `json:"name"`

	Description string `json:"description,omitempty"`

	User User `json:"user"`

	Images []Image `json:"images,omitempty"`
}