# internal/model/album_type.go
/*
 * Otto image service
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type AlbumType string

// List of AlbumType
const (
	ProfilePics AlbumType = "profile_pics"
	UserCreated AlbumType = "user_created"
	Livestream  AlbumType = "livestream"
)