# internal/model/image.go
/*
 * Otto image service
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

import (
	"time"
)

type Image struct {

	Uuid string `json:"uuid"`

	Filename string `json:"filename"`

	S3Key string `json:"s3_key"`

	CreatedAt time.Time `json:"created_at"`

	Link string `json:"link"`

	Album Album `json:"album"`

	User User `json:"user"`
}