# internal/model/new_album.go
/*
 * Otto image service
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

import (
	"encoding/json"
	"net/http"
)

type NewAlbum struct {

	Visibility Visibility `json:"visibility"`

	Name string `json:"name"`

	Description string `json:"description,omitempty"`

	Link string `json:"link,omitempty"`

	User User `json:"user,omitempty"`
}

func DecodeRequestToNewAlbum(r *http.Request) *NewAlbum {
	decoder := json.NewDecoder(r.Body)
	var data *NewAlbum
	err := decoder.Decode(&data)
	if err != nil {
		panic(err)
	}
	return data
}