# internal/model/new_image.go
/*
 * Otto image service
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

import (
	"encoding/json"
	"net/http"
)

type NewImage struct {

	Filename string `json:"filename"`

	Link string `json:"link,omitempty"`

	Image string `json:"image,omitempty"`
}

func DecodeRequestToNewImage(r *http.Request) *NewImage {
	decoder := json.NewDecoder(r.Body)
	var data *NewImage
	err := decoder.Decode(&data)
	if err != nil {
		panic(err)
	}
	return data
}