# internal/model/user.go
/*
 * Otto user service
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

import (
	"encoding/json"
)

// User struct for User
type User struct {

	Uuid string `json:"uuid"`

	Username string `json:"username"`

	Albums []Album `json:"albums,omitempty"`
}

func DecodeMessageToUser(message []byte) (*User, error) {
	user := &User{}
	err := json.Unmarshal(message, user)
	return user, err
}